cr feedback; move test to new package, update DataRegionTable
